Optimisations
A number of optimisations that make LOFTEE play nicely with
the modern ensembl-vep version of VEP. Overall runtime on a
test set of data is ~3x faster.

In a number of places in-memory copies of very large hashes
and arrays are being made when simply referring to the data
using the arrow notation on the reference would suffice.
There are likely more that I haven't replaced but the worst
offenders from an NYTProf profile have been changed.

Also replaced numerous uses of $transcript->get_all_Exons() and
get_all_Introns() with the VEP and cache-friendly
$transcript_variation->_exons() and _introns()

Similarly replaced the estimated translateable seq length method
with a fully accurate cache-friendly call that will be no slower.